Backend/requirements.txt
altair==5.0.1
attrs==24.2.0
blinker==1.6.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
Flask==2.2.5
Flask-Cors==5.0.0
gunicorn
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.3
packaging==23.2
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
protobuf==4.24.4
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b1
Pygments==2.17.2
Pympler==1.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2024.2
requests==2.31.0
rich==13.8.1
six==1.17.0
smmap==5.0.2
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.3.2
typing_extensions==4.7.1
urllib3==2.0.7
validators==0.20.0
Wand==0.6.13
watchdog==3.0.0
Werkzeug==2.2.3
zipp==3.15.0
